pacman verde

+++ Object.h
@@ -5,10 +5,10 @@
 #include "Vector.h"
 
 typedef struct color{
-	Uint8 r;
-	Uint8 g;
-	Uint8 b;
-	Uint8 a;
+	int r;
+	int g;
+	int b;
+	int a;
 } Color;
 
 typedef struct object{
